Add files via upload
In this project, we utilized the IRIS open-source dataset to demonstrate machine learning techniques. Specifically, we employed two conventional ML methods: RandomForest classifier and SVM (svc). These approaches yielded impressive results with classification accuracies of 96.66% and 100%, respectively. The code is designed to be straightforward and easily comprehensible, making it accessible for learning and experimentation.